+++ src/components/NetworkLogger.tsx
@@ -1,14 +1,17 @@
 import React, { useEffect, useState, useCallback } from 'react';
-import { Alert, View, StyleSheet, BackHandler } from 'react-native';
+import { Alert, View, StyleSheet, BackHandler, Share } from 'react-native';
 import logger from '../loggerSingleton';
 import NetworkRequestInfo from '../NetworkRequestInfo';
-import { ThemeContext, ThemeName } from '../theme';
+import { Theme, ThemeContext, ThemeName } from '../theme';
+import { setBackHandler } from '../backHandler';
+import { DeepPartial } from '../types';
 import RequestList from './RequestList';
 import RequestDetails from './RequestDetails';
-import { setBackHandler } from '../backHandler';
+import createHar from '../utils/createHar';
+import Unmounted from './Unmounted';
 
 interface Props {
-  theme?: ThemeName;
+  theme?: ThemeName | DeepPartial<Theme>;
   sort?: 'asc' | 'desc';
 }
 
@@ -25,6 +28,7 @@ const NetworkLogger: React.FC<Props> = ({ theme = 'light', sort = 'desc' }) => {
   );
   const [request, setRequest] = useState<NetworkRequestInfo>();
   const [showDetails, _setShowDetails] = useState(false);
+  const [mounted, setMounted] = useState(false);
 
   const setShowDetails = useCallback((shouldShow: boolean) => {
     _setShowDetails(shouldShow);
@@ -42,6 +46,7 @@ const NetworkLogger: React.FC<Props> = ({ theme = 'light', sort = 'desc' }) => {
     });
 
     logger.enableXHRInterception();
+    setMounted(true);
 
     return () => {
       // no-op if component is unmounted
@@ -68,12 +73,24 @@ const NetworkLogger: React.FC<Props> = ({ theme = 'light', sort = 'desc' }) => {
     return () => backHandler.remove();
   }, [showDetails, setShowDetails]);
 
+  const getHar = async () => {
+    const har = await createHar(logger.getRequests());
+
+    Share.share({
+      message: JSON.stringify(har),
+    });
+  };
+
   const showMore = () => {
     Alert.alert('More Options', undefined, [
       {
         text: 'Clear Logs',
         onPress: () => logger.clearRequests(),
         style: 'destructive',
+      },
+      {
+        text: 'Export all Logs',
+        onPress: getHar,
       },
       { text: 'Cancel', style: 'cancel' },
     ]);
@@ -91,15 +108,19 @@ const NetworkLogger: React.FC<Props> = ({ theme = 'light', sort = 'desc' }) => {
           </View>
         )}
         <View style={showDetails && !!request ? styles.hidden : styles.visible}>
-          <RequestList
-            requests={requests}
-            onShowMore={showMore}
-            showDetails={showDetails && !!request}
-            onPressItem={(item) => {
-              setRequest(item);
-              setShowDetails(true);
-            }}
-          />
+          {mounted && !logger.enabled && !requests.length ? (
+            <Unmounted />
+          ) : (
+            <RequestList
+              requests={requests}
+              onShowMore={showMore}
+              showDetails={showDetails && !!request}
+              onPressItem={(item) => {
+                setRequest(item);
+                setShowDetails(true);
+              }}
+            />
+          )}
         </View>
       </View>
     </ThemeContext.Provider>